backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
django-filter==23.5
celery==5.3.4
redis==5.0.1
python-decouple==3.8
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.34
geopy==2.4.1
requests==2.31.0
